apply of_type error message to load.options() as well
Improved the error reporting when linking strategy options from a base
class to another attribute that's off a subclass, where ``of_type()``
should be used. Previously, when :meth:`.Load.options` is used, the message
would lack informative detail that ``of_type()`` should be used, which was
not the case when linking the options directly. The informative detail now
emits even if :meth:`.Load.options` is used.

Fixes: #9182
Change-Id: Ibc14923d0cbca9114316cb7db2b30f091dc28af8